[#29] Rename package module > modules, model > models
