+++ shared/src/commonMain/kotlin/co/nimblehq/kmm/template/data/remote/datasources/RemoteDataSource.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.kmm.template.data.remote.datasources
 
-import co.nimblehq.kmm.template.data.remote.model.responses.Response
+import co.nimblehq.kmm.template.data.remote.models.responses.Response
 import co.nimblehq.kmm.template.data.remote.ApiClient
 
 interface RemoteDataSource {
