+++ shared/src/commonMain/kotlin/co/nimblehq/kmm/template/data/remote/models/responses/Response.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.kmm.template.data.remote.model.responses
+package co.nimblehq.kmm.template.data.remote.models.responses
 
 import co.nimblehq.kmm.template.domain.models.Model
 import kotlinx.serialization.SerialName
