+++ shared/src/commonMain/kotlin/co/nimblehq/kmm/template/data/repositories/RepositoryImpl.kt
@@ -2,7 +2,7 @@ package co.nimblehq.kmm.template.data.repositories
 
 import co.nimblehq.kmm.template.data.extensions.flowTransform
 import co.nimblehq.kmm.template.data.remote.datasources.RemoteDataSource
-import co.nimblehq.kmm.template.data.remote.model.responses.toModel
+import co.nimblehq.kmm.template.data.remote.models.responses.toModel
 import co.nimblehq.kmm.template.domain.models.Model
 import co.nimblehq.kmm.template.domain.repositories.Repository
 import kotlinx.coroutines.flow.Flow
