+++ shared/src/commonMain/kotlin/co/nimblehq/kmm/template/di/Koin.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.kmm.template.di
 
-import co.nimblehq.kmm.template.di.module.*
+import co.nimblehq.kmm.template.di.modules.*
 import org.koin.core.KoinApplication
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration
